moved app logic from index.ts to app.ts file
